+++ app/src/main/java/com/example/mvishowcase/feature/home/di/HomeContainer.kt
@@ -3,19 +3,19 @@ package com.example.mvishowcase.feature.home.di
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.example.mvishowcase.domain.repository.CountryRepository
-import com.example.mvishowcase.domain.usecase.GetCountriesUseCase
+import com.example.mvishowcase.domain.usecase.SearchCountriesUseCase
 import com.example.mvishowcase.feature.home.presentation.HomeViewModel
 
 class HomeContainer(private val countryRepository: CountryRepository) {
 
-    private val getCountriesUseCase: GetCountriesUseCase by lazy {
-        GetCountriesUseCase(countryRepository)
+    private val searchCountriesUseCase: SearchCountriesUseCase by lazy {
+        SearchCountriesUseCase(countryRepository)
     }
 
     val viewModelFactory: ViewModelProvider.Factory = object : ViewModelProvider.Factory {
         @Suppress("UNCHECKED_CAST")
         override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            return HomeViewModel(getCountriesUseCase) as T
+            return HomeViewModel(searchCountriesUseCase) as T
         }
     }
 }