+++ app/src/main/java/com/example/mvishowcase/feature/home/presentation/HomeUiState.kt
@@ -1,0 +1,6 @@
+package com.example.mvishowcase.feature.home.presentation
+sealed interface HomeUiState {
+    data object Loading : HomeUiState
+    data class Error(val message: String) : HomeUiState
+    data object Success : HomeUiState
+}